Change light duration variables to uint16_t
Changed duration variables from float to uint16_t and updated print format.

+++ ECE 455/Project1_Code/src/main.c
@@ -223,9 +223,9 @@ static void traffic_gen_task ( void *pvParameters ) {
 }
 
 static void light_state_task ( void *pvParameters ) {
-    float r_dur = 3000.0; // assume milliseconds
-    float y_dur = 3000.0; // assume milliseconds
-    float g_dur = 3000.0; // assume milliseconds
+    uint16_t r_dur = 3000.0; // assume milliseconds
+    uint16_t y_dur = 3000.0; // assume milliseconds
+    uint16_t g_dur = 3000.0; // assume milliseconds
     
     uint16_t ADC_val = 0;
     uint16_t rx_data;
@@ -239,9 +239,9 @@ static void light_state_task ( void *pvParameters ) {
                 {
                     r_dur = 9999.0 * (2 - ADC_val/ADC_MAX);
                     g_dur = 9999.0 * (1 + ADC_val/ADC_MAX);
-                	printf("r_dur %.5f.\n", r_dur);
-                	printf("y_dur %.5f.\n", y_dur);
-                	printf("g_dur %.5f.\n", g_dur);
+                	printf("r_dur %u.\n", r_dur);
+                	printf("y_dur %u.\n", y_dur);
+                	printf("g_dur %u.\n", g_dur);
                 }
             }
